requirements.txt
pandas==1.3.5
numpy==1.20.1
selenium==3.141.0
beautifulsoup4==4.9.3
webdriver_manager==4.0.1
google-cloud-storage==2.10.0